+++ robocluster/util.py
@@ -1,9 +1,19 @@
 """Utility functions for robocluster."""
 
 import re
+import socket
+import ipaddress
+import struct
 from functools import wraps
 from inspect import iscoroutinefunction
 
+def ip_info(addr):
+    """Verify and detecet ip address family."""
+    addr = ipaddress.ip_address(addr)
+    if isinstance(addr, ipaddress.IPv6Address):
+        return socket.AF_INET6, addr
+    else:
+        return socket.AF_INET, addr
 
 def as_coroutine(func):
     """
